NiceGUI - MyApp/app/requirements.txt
loguru
nicegui
pytest
pytest-asyncio

pytest-playwright
